api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["codenet_toanlk <toanlk@codenet.jp>"]
packages = [
    { include = "__init__.py", from = "tasks" },
]

[tool.poetry.scripts]
task = "__init__:cli"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
mangum = "^0.11.0"
python-dotenv = "^0.17.1"
peewee = "^3.14.4"
psycopg2-binary = "^2.8.6"
python-jose = "^3.2.0"
redis = "^3.5.3"
requests = "^2.25.1"
jwcrypto = "^0.8"
mecab-python3 = "^1.0.3"
jaconv = "^0.3"
unidic-lite = "^1.0.8"
python-i18n = {extras = ["YAML"], version = "^0.3.9"}
PyYAML = "^5.4.1"
boto3 = "^1.17.84"
pydash = "^5.0.0"
click = "^7.1"
pip = "^22.0.4"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.8.2"
black = "^21.5b0"
flake8 = "^3.9.1"
pytest-cov = "^2.12.0"

[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
